% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catage.R
\docType{data}
\name{catage}
\alias{catage}
\alias{catage.long}
\alias{catage.taf}
\alias{catage.xtab}
\title{Catch at Age Examples}
\format{All three data frames contain the same data in alternative formats:
\tabular{lll}{
  \code{catage.long} \tab long format     \tab (three columns)\cr
  \code{catage.taf}  \tab TAF format      \tab (year column)\cr
  \code{catage.xtab} \tab crosstab format \tab (year as row names)
}}
\description{
Small catch-at-age table to describe three formats to store year-age values
in a data frame.
}
\references{
ICES. 2016. Report of the working group on the assessment of demersal stocks
in the North Sea and Skagerrak (WGNSSK).
\href{http://ices.dk/sites/pub/Publication\%20Reports/Expert\%20Group\%20Report/acom/2016/WGNSSK/01\%20WGNSSK\%20report\%202016.pdf}{\cite{ICES
CM 2016/ACOM:14}}, p. 656.
}
