% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.boot.entries.R
\name{check.boot.data}
\alias{check.boot.data}
\title{Check boot data entries}
\usage{
check.boot.data()
}
\value{
logical vector indicating which entries are present (TRUE) or missing (FALSE)
}
\description{
Check that all data entries in boot/DATA.bib are present in the boot/data folder
}
