% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.R
\name{cp}
\alias{cp}
\title{Copy Files}
\usage{
cp(from, to)
}
\arguments{
\item{from}{source filenames, e.g. \code{*.csv}.}

\item{to}{destination filenames, or directory.}
}
\value{
\code{TRUE} for success, \code{FALSE} for failure, invisibly.
}
\description{
Copy files, overwriting existing files if necessary, and returning the result
invisibly.
}
\examples{
\dontrun{
write(pi, "A.txt")
cp("A.txt", "B.txt")

## Copy directory tree
cp(system.file(package="datasets"), ".")
mkdir("everything")
cp("datasets/*", "everything")
}

}
\seealso{
\code{\link{file.copy}} is the base function to copy files.

\code{\link{icesTAF-package}} gives an overview of the package.
}
