% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach.packages.R
\name{detach.packages}
\alias{detach.packages}
\title{Detach Packages}
\usage{
detach.packages(quiet = FALSE)
}
\arguments{
\item{quiet}{whether to suppress messages.}
}
\description{
Detach all non-base packages that have been attached using \code{library} or
\code{taf.library}.
}
\examples{
\dontrun{
detach.packages()
}

}
\seealso{
\code{\link{detach}} is the underlying base function to detach a package.

\code{\link{taf.library}} loads a package from \verb{bootstrap/library}.

\code{\link{icesTAF-package}} gives an overview of the package.
}
