% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.tree.R
\name{dir.tree}
\alias{dir.tree}
\title{Print a directory tree}
\usage{
dir.tree(path = ".")
}
\arguments{
\item{path}{the directory for which the listing is to be shown}
}
\description{
Print the directory tree and file contents in a pretty way
}
\examples{
\dontrun{

library(icesTAF)

# Download a TAF analysis
dir.tree()
}

}
\seealso{
\link{list.files}
}
