% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos2unix.R, R/unix2dos.R
\name{dos2unix}
\alias{dos2unix}
\alias{unix2dos}
\title{Convert Line Endings}
\usage{
dos2unix(file)

unix2dos(file)
}
\arguments{
\item{file}{a filename.}
}
\description{
Convert line endings in a text file between Dos (CRLF) and Unix (LF) format.
}
\note{
TAF uses \code{unix2dos} to ensure that text files on the FTP server have Dos
line endings.
}
\examples{
\dontrun{
write("123", "test.txt")
dos2unix("test.txt")  # file is now 4 bytes
unix2dos("test.txt")  # file is now 5 bytes
}

}
\seealso{
\code{\link{file}} is used to open a binary connection to the text file.

\code{\link{writeLines}} is used to apply CRLF or LF line endings.

\code{\link{icesTAF-package}} gives an overview of the package.
}

