% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.analysis.R
\name{download.analysis}
\alias{download.analysis}
\title{Download a TAF analysis}
\usage{
download.analysis(repo, dir = tempdir())
}
\arguments{
\item{repo}{The full name of the GitHub repository, e.g. "ices-taf/2019_san.sa.6".}

\item{dir}{the directory to place the TAF project}
}
\description{
Download the code for a TAF analysis from GitHub.
}
\examples{
\dontrun{

library(icesTAF)

# Download a TAF analysis
run_dir <- download.analysis("ices-taf/2019_san.sa.6", dir = ".")

# run the analysis
run.analysis(run_dir)
}

}
\seealso{
\link{run.analysis}
}
