% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.github.R
\name{download.github}
\alias{download.github}
\title{Download GitHub Repository}
\usage{
download.github(repo, dir = ".")
}
\arguments{
\item{repo}{GitHub reference of the form \verb{owner/repo[/subdir]@ref}.}

\item{dir}{directory to download to.}
}
\description{
Download a repository from GitHub in \file{tar.gz} format.
}
\note{
In general, TAF scripts do not access the internet using
\code{download.github} or similar functions. Instead, data and software are
declared in \file{DATA.bib} and \file{SOFTWARE.bib} and then downloaded using
\code{\link{taf.bootstrap}}. The exception is when a bootstrap data script is
used to fetch data files from a web service (see \code{\link{process.bib}}).
}
\examples{
\dontrun{
download.github("ices-tools-prod/icesTAF@2.0-0")
}

}
\seealso{
\code{\link{download}} downloads a file.

\code{\link{untar}} extracts a \verb{tar.gz} archive.

\code{\link{icesTAF-package}} gives an overview of the package.
}
