% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flr2taf.R
\name{flr2taf}
\alias{flr2taf}
\title{Convert FLR Table to TAF Format}
\usage{
flr2taf(x, year = TRUE, na = NULL)
}
\arguments{
\item{x}{a table of class \code{FLQuant}.}

\item{year}{whether the year should be stored in the first column, instead of
row names.}

\item{na}{a value that should be converted to NA.}
}
\value{
A data frame in TAF format.
}
\description{
Convert a table from FLR to TAF format.
}
\note{
FLR uses the \code{FLQuant} class to store tables as 6-dimensional arrays,
while TAF tables are stored as data frames with a year column.
}
\examples{
x <- array(round(runif(70)/3, 1), dim=c(7,10,1,1,1,1))
dimnames(x) <- list(age=2:8, year=1991:2000,
                    unit="unique", season="all", area="unique", iter=1)
flr2taf(x)
flr2taf(x, year=FALSE, na=0)

}
\seealso{
\code{\link{as.data.frame}} is a method provided by the \pkg{FLCore} package
to convert \code{FLQuant} tables to a 7-column long format.

\code{\link{taf2long}} converts TAF tables to long format.

\code{\link{icesTAF-package}} gives an overview of the package.
}
