% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesTAF-import.R
\name{icesTAF-import}
\alias{icesTAF-import}
\alias{already.in.taf.library}
\alias{clean}
\alias{clean.data}
\alias{clean.library}
\alias{clean.software}
\alias{convert.spaces}
\alias{cp}
\alias{deps}
\alias{detach.packages}
\alias{div}
\alias{dos2unix}
\alias{download}
\alias{download.github}
\alias{draft.data}
\alias{draft.software}
\alias{ds.file}
\alias{ds.package}
\alias{extract.subdir}
\alias{file.encoding}
\alias{flr2taf}
\alias{get.remote.sha}
\alias{is.r.package}
\alias{latin1.to.utf8}
\alias{lim}
\alias{line.endings}
\alias{long2taf}
\alias{long2xtab}
\alias{make}
\alias{makeAll}
\alias{makeTAF}
\alias{mkdir}
\alias{msg}
\alias{os}
\alias{os.linux}
\alias{os.macos}
\alias{os.unix}
\alias{os.windows}
\alias{parse.repo}
\alias{period}
\alias{plus}
\alias{process.bibfile}
\alias{process.entry}
\alias{read.bib}
\alias{read.taf}
\alias{rmdir}
\alias{rnd}
\alias{sam2taf}
\alias{sourceAll}
\alias{sourceDir}
\alias{sourceTAF}
\alias{stamp.description}
\alias{taf.boot.path}
\alias{taf.bootstrap}
\alias{taf.data.path}
\alias{taf.install}
\alias{taf.libPaths}
\alias{taf.library}
\alias{taf.png}
\alias{taf.session}
\alias{taf.skeleton}
\alias{taf.sources}
\alias{taf.unzip}
\alias{taf2long}
\alias{taf2xtab}
\alias{tafpng}
\alias{tt}
\alias{unix2dos}
\alias{utf8.to.latin1}
\alias{write.taf}
\alias{xtab2long}
\alias{xtab2taf}
\alias{zoom}
\alias{zoom.trellis}
\title{-}
\description{
All TAF functions are imported into icesTAF to allow indirect namespace calls
such as \code{icesTAF::read.taf}, which can be beneficial. The direct
\code{TAF::read.taf} call also works.

This comes with the minor cost that two help pages exist. The actual user
documentation of imported TAF functions are found in the TAF package.
}
\seealso{
\code{\link{TAF-package}} provides help pages for functions defined in the
\pkg{TAF} package.

\code{\link{icesTAF-package}} gives an overview of the \pkg{icesTAF} package.
}
