% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesTAF-internal.R, R/taf.roxygenise.R
\name{icesTAF-internal}
\alias{icesTAF-internal}
\alias{roxy_tag_parse.roxy_tag_tafKey}
\alias{roxy_tag_parse.roxy_tag_tafOriginator}
\alias{roxy_tag_parse.roxy_tag_tafYear}
\alias{roxy_tag_parse.roxy_tag_tafPeriod}
\alias{roxy_tag_parse.roxy_tag_tafAccess}
\alias{roxy_tag_parse.roxy_tag_tafSource}
\title{Various TAF Helper Functions}
\usage{
\method{roxy_tag_parse}{roxy_tag_tafKey}(x)

\method{roxy_tag_parse}{roxy_tag_tafOriginator}(x)

\method{roxy_tag_parse}{roxy_tag_tafYear}(x)

\method{roxy_tag_parse}{roxy_tag_tafPeriod}(x)

\method{roxy_tag_parse}{roxy_tag_tafAccess}(x)

\method{roxy_tag_parse}{roxy_tag_tafSource}(x)
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{icesTAF-package}} gives an overview of the package.
}
\keyword{internal}
