% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesTAF-package.R
\docType{package}
\name{icesTAF-package}
\alias{icesTAF}
\alias{icesTAF-package}
\title{Helper Functions for the ICES Transparent Assessment Framework}
\description{
Helper functions for the Transparent Assessment Framework (TAF) used by the
International Council for the Exploration of the Sea (ICES).
}
\details{
\emph{Data-limited stocks:}
\tabular{ll}{
  \code{\link{readDLS}}  \tab read \code{DLS3.2} output from file\cr
  \code{\link{writeDLS}} \tab write \code{DLS3.2} output to file
}
}
\author{
Arni Magnusson and Colin Millar.
}
\references{
ICES Transparent Assessment Framework: \url{http://taf.ices.dk}.
}

