% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ido.R
\name{ido}
\alias{ido}
\title{Invisible Dev Off}
\usage{
ido(...)
}
\arguments{
\item{\dots}{passed to \code{dev.off}.}
}
\description{
Close graphics device quietly.
}
\note{
A simple convenience function to shorten
\preformatted{invisible(dev.off())}
to
\preformatted{ido()}
}
\examples{
\dontrun{
tafpng("myplot")
plot(1)
ido()
}

}
\seealso{
\code{\link{dev.off}} is the underlying function used to close a graphics
device.

\code{\link{tafpng}} opens a PNG graphics device.

\code{\link{icesTAF-package}} gives an overview of the package.
}
