% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.deps.R
\name{install.deps}
\alias{install.deps}
\title{Install packages dependencies of a TAF analysis}
\usage{
install.deps(path = ".", ...)
}
\arguments{
\item{path}{a directory or file containing R scripts.}

\item{...}{arguments passed on to \link{install.packages}}
}
\description{
Search R scripts for packages that are required and install them.
}
\examples{
\dontrun{

library(TAF)

# Download a TAF analysis
download("https://github.com/ices-taf/2019_san.sa.6/archive/refs/heads/master.zip")
unzip("master.zip")

# move into analysis folder
setwd("2019_san.sa.6-master")

# list dependencies
deps()

# install dependencies
install.deps()
}

}
\seealso{
\link{deps}
}
