% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin1.to.utf8.R
\name{latin1.to.utf8}
\alias{latin1.to.utf8}
\title{Convert Latin 1 to UTF-8}
\usage{
latin1.to.utf8(file, force = FALSE)
}
\arguments{
\item{file}{filename of a text file, e.g. source code or data.}

\item{force}{whether to perform the conversion even when it's not clear that
the file is currently \code{"latin1"} encoded. Not recommended.}
}
\description{
Convert file encoding from \verb{"latin1"} to \verb{"UTF-8"}.
}
\details{
This function should only be used on a file that is currently \code{"latin1"}
encoded. The default behavior is to verify this with
\code{\link{enc.latin1}}, but this check can be suppressed using
\code{force = TRUE}.
}
\note{
For robust and efficient batch conversion of multiple files, the
\command{iconv} shell command is more appropriate.
}
\seealso{
\code{\link{Encoding}} examines the encoding of a string.

\code{\link{enc}} examines the encoding of a file.

\code{\link{icesTAF-package}} gives an overview of the package.
}
