% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lim.R
\name{lim}
\alias{lim}
\title{Axis Limits}
\usage{
lim(x, mult = 1.1)
}
\arguments{
\item{x}{a vector of data values.}

\item{mult}{a number to multiply with the highest data value.}
}
\value{
A vector of length two, which can be used as axis limits.
}
\description{
Compute axis limits. The lower limit is 0 and the upper limit is determined
by the highest data value, times a multiplier.
}
\examples{
plot(rivers, ylim=lim(rivers))

}
\seealso{
\code{\link{icesTAF-package}} gives an overview of the package.
}
