% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.endings.R
\name{line.endings}
\alias{line.endings}
\title{Line Endings}
\usage{
line.endings(file)
}
\arguments{
\item{file}{a filename.}
}
\value{
String indicating the line endings: \code{"Dos"} or \code{"Unix"}.
}
\description{
Examine whether file has Dos or Unix line endings.
}
\examples{
\dontrun{
file <- system.file(package="icesTAF", "DESCRIPTION")
line.endings(file)
}

}
\seealso{
\code{\link{file.encoding}} examines the encoding of a file.

\code{\link{dos2unix}} and \code{\link{unix2dos}} convert line endings.

\code{\link{icesTAF-package}} gives an overview of the package.
}
