% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2xtab.R
\name{long2xtab}
\alias{long2xtab}
\title{Convert Long Table to Crosstab Format}
\usage{
long2xtab(x)
}
\arguments{
\item{x}{a data frame in long format.}
}
\value{
A data frame with years as row names.
}
\description{
Convert a table from long format to crosstab format.
}
\examples{
long2xtab(catage.long)

}
\seealso{
\code{\link{catage.long}} and \code{\link{catage.xtab}} describe the long and
crosstab formats.

\code{\link{long2taf}} and \code{\link{taf2xtab}} are the underlying
functions that perform the conversion.

\code{\link{icesTAF-package}} gives an overview of the package.
}
