% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAll.R
\name{makeAll}
\alias{makeAll}
\title{Run All TAF Scripts as Needed}
\usage{
makeAll(path = ".", quiet = TRUE, debug = FALSE)
}
\arguments{
\item{path}{directory containing TAF scripts.}

\item{quiet}{whether to suppress messages when nothing is done.}

\item{debug}{whether to show a diagnostic table of files and time last
modified.}
}
\value{
Logical vector indicating which scripts were run.
}
\description{
Run TAF scripts that have changed, or if previous steps were rerun.
}
\examples{
\dontrun{
makeAll()
}

}
\seealso{
\code{\link{make}} runs a TAF script if needed.

\code{\link{sourceAll}} runs all TAF scripts in a directory.

\code{\link{icesTAF-package}} gives an overview of the package.
}
