% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Show Message}
\usage{
msg(...)
}
\arguments{
\item{\dots}{passed to \code{message}.}
}
\description{
Show a message, as well as the current time.
}
\examples{
msg("Running script.R")

}
\seealso{
\code{\link{message}} is the base function to show messages, without the
current time.

\code{\link{sourceTAF}} reports progress using \code{msg}.

\code{\link{icesTAF-package}} gives an overview of the package.
}
