% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.unix.R, R/os.windows.R
\name{os.unix}
\alias{os.unix}
\alias{os.windows}
\title{OS Family}
\usage{
os.unix()

os.windows()
}
\description{
Determine whether operating system is Windows or Unix-compatible.
}
\note{
These shorthand functions can be useful when writing workaround solutions in
platform-independent scripts.
}
\examples{
os.unix()
os.windows()

}
\seealso{
\code{\link{.Platform}} reports the \code{OS.type} family.

\code{\link{icesTAF-package}} gives an overview of the package.
}
