% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period}
\alias{period}
\title{Period}
\usage{
period(x, y = NULL)
}
\arguments{
\item{x}{the first year, a vector of years, or a data frame containing
a \code{Year} column.}

\item{y}{the last year, if \code{x} is only the first year.}
}
\value{
A string of the form \code{"1990-2000"}.
}
\description{
Paste two years to form a \code{period} string.
}
\details{
If \code{x} is a vector or a data frame, the lowest and highest years are
used, and \code{y} is ignored.
}
\note{
This function can be useful when working with \code{\link{draft.data}}.
}
\examples{
period(1963, 1970)
period(range(catage.taf$Year))
period(catage.taf$Year)
period(catage.taf)

}
\seealso{
\code{\link{paste}} is the underlying function to paste strings.

\code{\link{draft.data}} has an argument called \code{period}.

\code{\link{icesTAF-package}} gives an overview of the package.
}
