% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.taf.R
\name{print.taf}
\alias{print.taf}
\title{Print TAF Table}
\usage{
\method{print}{taf}(x, quote = FALSE, na.print = "", right = TRUE,
  row.names = FALSE, ...)
}
\arguments{
\item{x}{a data frame in TAF format.}

\item{quote}{whether all values should be quoted.}

\item{na.print}{string to use for missing values. Alternatively, a logical
where \code{TRUE} is interpreted as \code{"NA"} and \code{FALSE} is
interpreted as \code{""}.}

\item{right}{whether strings should be right aligned.}

\item{row.names}{whether to include row names.}

\item{\dots}{passed to \code{print.default}.}
}
\description{
Print a TAF table, controlling how \code{NA} values are shown.
}
\examples{
x <- data.frame(Year=c(2000,NA), SSB=c(123456,800), Fbar=c(1.234,1.5))

print.taf(x)
print.taf(x, na="-")
print.taf(x, na=TRUE)

}
\seealso{
\code{\link{print.default}} is the underlying function used to print a
matrix.

\code{\link{write.taf}} writes a TAF table to a file.

\code{\link{icesTAF-package}} gives an overview of the package.
}
