% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.bib.R
\name{process.bib}
\alias{process.bib}
\title{Process Metadata}
\usage{
process.bib(bibfile)
}
\arguments{
\item{bibfile}{metadata file to process, either \code{"DATA.bib"} or
\code{"SOFTWARE.bib"}.}
}
\description{
Read and process metadata entries in a \verb{*.bib} file.
}
\note{
This is a helper function for \code{\link{taf.bootstrap}} and can also be
used in a custom \verb{bootstrap.R} script. It should be called within the
\file{bootstrap} directory that contains the metadata file.

Within each metadata entry, the \dfn{source} field specifies where data or
software originate from. There are four types of values that can be used in
the source field:
\enumerate{
\item GitHub reference of the form \verb{owner/repo[/subdir]@ref},
identifying a specific version of an R package.
\item URL starting with \verb{http} or \verb{https}, identifying a file to
download.
\item Filename, identifying the location of a file provided by the user.
\item Special value \code{file}, indicating that the metadata key points to a
file location.
}

Consider, for example, the following metadata entry from a \verb{DATA.bib}
file:
\preformatted{@Misc{catch.csv,
  originator = {WGEF},
  year       = {2015},
  title      = {Annual catch of rjm-347d},
  period     = {2012-2014},
  source     = {file},
}}
Here, a data file is described using the \verb{@Misc} entry type and the
string following the entry type is called a \dfn{key}. The next fields state
that this file was supplied by the Elasmobranch working group in 2015 and it
contains the annual catch of North Sea spotted ray from 2012 to 2014. The
special value \verb{source = {file}} means that the key, in this case
\verb{catch.csv}, is the name of the file located inside
\verb{bootstrap/initial/data}. This \verb{file} shorthand notation is
equivalent to specifying the relative path:
\verb{source = {initial/data/catch.csv}}.

Another example metadata entry is from a \verb{SOFTWARE.bib} file:
\preformatted{@Manual{icesAdvice,
  author  = {Arni Magnusson and Colin Millar and Anne Cooper},
  year    = {2018},
  title   = {icesAdvice: Functions Related to ICES Advice},
  version = {2.0-0, released 2018-12-07},
  source  = {cran/icesAdvice@2.0-0},
}}
This entry describes a specific version of an R package that is required for
the TAF analysis. It is similar, but not identical, to the output from the R
command \verb{citation("icesAdvice")}. The version field specifies the
version number and release date, with a corresponding GitHub reference. When
an R package is not an official release but a development version, the
version and source may look like this,
\preformatted{  version = {2.6.6, committed 2018-02-21},
  source  = {flr/FLCore@d0333c1},}
or this:
\preformatted{  version = {0.5.4 components branch, committed 2018-03-12},
  source  = {fishfollower/SAM/stockassessment@25b3591},}
For development versions like these, the version number itself may not be
important or accurate, but the branch name may be informative, and the
7-character SHA reference code is a pointer to the exact version of the
package required for the analysis.

In summary, the metadata are similar to bibliographic entries, with the
important addition of source directives that guide the bootstrap procedure to
set up data files and software.
}
\examples{
\dontrun{
process.bib("DATA.bib")
process.bib("SOFTWARE.bib")
}

}
\seealso{
\code{\link{taf.bootstrap}} calls \code{process.bib} to process metadata.

\code{\link{icesTAF-package}} gives an overview of the package.
}
