% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.artifacts.R
\name{read.artifacts}
\alias{read.artifacts}
\title{Read an Artifact File}
\usage{
read.artifacts(file = "ARTIFACTS.json", check = TRUE)
}
\arguments{
\item{file}{the name of the artifacts file to read}

\item{check}{logical, should the artifact be checked for validity?}
}
\description{
Read and validate an artifact file.
}
