% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bib.R
\name{read.bib}
\alias{read.bib}
\title{bibtex parser}
\usage{
read.bib(file)
}
\arguments{
\item{file}{bib file to parse}
}
\value{
TRUE if successfull, FALSE otherwise
}
\description{
Parser for bibliography databases written in the bib format.
}
\details{
Upload a table of F at age to the ICES TAF results database.
The \code{data} and \code{assessment_info} will be checked against
a schema, and any errors reported back to the user as attributes to
the FALSE return value.
}
\note{
The \code{data} argument expects a data.frame with the first
column named "year" as described in the help for
\code{\link{xtab2taf}}.
}
\seealso{
The \code{\link{xtab2taf}} function is to create a data.frame

\code{\link{taf.png}} opens a PNG graphics device.

\code{\link{icesTAF-package}} gives an overview of the package.
}
