% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDLS.R
\name{readDLS}
\alias{readDLS}
\title{Read DLS Advice from File}
\usage{
readDLS(file)
}
\arguments{
\item{file}{a filename.}
}
\value{
A list containing \code{advice} and other elements showing intermediate steps
in the calculations.
}
\description{
Read the advisory calculations for a data-limited stock (DLS) from a file.
}
\examples{
\dontrun{
library(icesAdvice)
survey <- data.frame(year=2001:2010, randu[1:10,])
dls <- DLS3.2(1000, survey$y)
writeDLS(dls, "dls-output.txt")
dls2 <- readDLS("dls-output.txt")
}
}
\seealso{
\code{\link{writeDLS}} writes DLS advisory calculations to a file.

\code{\link[icesAdvice]{DLS3.2}} in the \pkg{icesAdvice} package can be used
to calculate catch advice for data-limited stocks.

\code{\link{icesTAF-package}} gives an overview of the package.
}

