% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdir.R
\name{rmdir}
\alias{rmdir}
\title{Remove Empty Directory}
\usage{
rmdir(path, recursive = FALSE)
}
\arguments{
\item{path}{a directory name.}

\item{recursive}{whether to remove empty subdirectories as well.}
}
\value{
\code{TRUE} for success, \code{FALSE} for failure, invisibly.
}
\description{
Remove empty directories under any operating system.
}
\note{
To tidy up a project directory, it can be useful to remove empty
subdirectories, while leaving non-empty directories intact.

The base function \code{unlink(dir, recursive=FALSE)} does not remove empty
directories in Windows and \code{unlink(dir, recursive=TRUE)} removes
non-empty directories, making it unsuitable for tidying up empty ones.
}
\examples{
\dontrun{
mkdir("emptydir")
rmdir("emptydir")

mkdir("outer/inner")
rmdir("outer/inner", recursive=TRUE)
}

}
\seealso{
\code{\link{unlink}} can remove non-empty directories.

\code{\link{mkdir}} and \code{\link{dir.create}} create an empty directory.

\code{\link{icesTAF-package}} gives an overview of the package.
}
