% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.analysis.R
\name{run.analysis}
\alias{run.analysis}
\title{Run a TAF analysis}
\usage{
run.analysis(dir)
}
\arguments{
\item{dir}{the directory where the TAF project is located}
}
\description{
Run the code for a TAF analysis locally.
}
\examples{
\dontrun{

library(icesTAF)

# Download a TAF analysis
run_dir <- download.analysis("ices-taf/2019_san.sa.6", dir = ".")

# run the analysis
run.analysis(run_dir)
}

}
\seealso{
\link{download.analysis}
}
