% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceAll.R
\name{sourceAll}
\alias{sourceAll}
\title{Run All TAF Scripts}
\usage{
sourceAll(path = ".", clean = TRUE, ...)
}
\arguments{
\item{path}{directory containing TAF scripts.}

\item{clean}{whether to remove the target directory before running each
script.}

\item{\dots}{passed to \code{\link{sourceTAF}}.}
}
\value{
Logical vector, indicating which scripts ran without errors.
}
\description{
Run all TAF scripts found in a directory.
}
\note{
TAF scripts that will be run if they exist: \code{data.R}, \code{input.R},
\code{model.R}, \code{output.R}, and \code{report.R}.
}
\examples{
\dontrun{
sourceAll()
}

}
\seealso{
\code{\link{sourceTAF}} runs a TAF script.

\code{\link{makeAll}} runs all TAF scripts as needed.

\code{\link{clean}} cleans TAF directories.

\code{\link{icesTAF-package}} gives an overview of the package.
}
