% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.taf.R
\name{t.taf}
\alias{t.taf}
\title{Convert TAF Table to Transposed Crosstab Format}
\usage{
\method{t}{taf}(x)
}
\arguments{
\item{x}{a data frame in TAF format.}
}
\value{
A data frame with years as column names.
}
\description{
Convert a table from TAF format to transposed crosstab format.
}
\note{
Transposing can be useful when comparing TAF tables to stock assessment
reports.
}
\examples{
taf2xtab(catage.taf)
t.taf(catage.taf)

}
\seealso{
\code{\link{catage.taf}} describes the TAF format.

\code{\link{taf2xtab}} converts a TAF table to crosstab format, without
transposing.

\code{\link{icesTAF-package}} gives an overview of the package.
}
