% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.boot.path.R
\name{taf.boot.path}
\alias{taf.boot.path}
\title{Construct Path to a TAF bootstrap folder}
\usage{
taf.boot.path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors. Long vectors are not supported.}

\item{fsep}{the path separator to use (assumed to be ASCII).}
}
\value{
character
}
\description{
Construct the path to a file in the TAF bootstrap data folder
from components in a platform-independent way.  This function
checks to see if R is running in the bootstrap folder - i.e.
`taf.bootstrap()` is running, and adjusts the path accordingly.
}
\details{
This function, simplifies the construction of file paths to
the boot (bootstrap) folder.
}
\seealso{
\link{file.path}
}
