% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.citations.R
\name{taf.citations}
\alias{taf.citations}
\title{TAF Citations}
\usage{
taf.citations(file)
}
\arguments{
\item{file}{data citations file.}
}
\value{
Data frame containing data filenames and description.
}
\description{
Generate BibTeX references, based on data citations file and TAF repository
metadata.
}
