% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.install.R
\name{taf.install}
\alias{taf.install}
\title{TAF Install}
\usage{
taf.install(targz, lib = "bootstrap/library", quiet = FALSE)
}
\arguments{
\item{targz}{package filename, see example.}

\item{lib}{location of local TAF library.}

\item{quiet}{whether to suppress messages.}
}
\description{
Install package in \file{tar.gz} format in local TAF library.
}
\note{
After installing the package, this function writes the remote SHA reference
code into the package files \verb{DESCRIPTION} and \verb{Meta/package.rds}.
}
\examples{
\dontrun{
taf.install("bootstrap/software/FLAssess_f1e5acb.tar.gz")
}

}
\seealso{
\code{\link{taf.bootstrap}} calls \code{taf.install} to install R packages,
via \code{\link{process.bib}}.

\code{\link{download.github}} downloads a GitHub repository, for example a
package.

\code{\link{install.packages}} is the underlying base function to install a
package.

\code{\link{icesTAF-package}} gives an overview of the package.
}
