% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.libPaths.R
\name{taf.libPaths}
\alias{taf.libPaths}
\title{Add TAF Library Path}
\usage{
taf.libPaths()
}
\value{
The resulting vector of file paths.
}
\description{
Add TAF library to the search path for R packages.
}
\note{
Specifically, this function sets \code{"bootstrap/library"} as the first
element of \code{.libPaths()}. This is rarely beneficial in TAF scripts, but
can be useful when using the \pkg{sessioninfo} package, for example.
}
\examples{
\donttest{
taf.libPaths()
}

}
\seealso{
\code{\link{.libPaths}} is the underlying function to modify the search path
for R packages.

\code{\link{taf.library}} loads a package from \verb{bootstrap/library}.

\code{\link{icesTAF-package}} gives an overview of the package.
}
