% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.library.R
\name{taf.library}
\alias{taf.library}
\title{Enable TAF Library}
\usage{
taf.library(names = FALSE)
}
\arguments{
\item{names}{whether to return the names of packages currently installed in
the TAF library.}
}
\value{
The resulting search path, or names of packages if \code{names = TRUE}.
}
\description{
Add local TAF library \file{bootstrap/library} to the search path, where
packages are stored.
}
\note{
This function inserts the entry \code{"bootstrap/library"} in front of the
existing library search path.

The purpose of the TAF library is to retain R packages used in a TAF
assessment that are not archived on CRAN, to support long-term
reproducibility of TAF assessments.
}
\examples{
\dontrun{
taf.library()
library(this)
library(that)

# Show packages in TAF library
taf.library(TRUE)
}

}
\seealso{
\code{\link{taf.install}} installs a package from GitHub into the TAF
library.

\code{\link{.libPaths}} is the underlying base function to set the library
search path.

\code{\link{icesTAF-package}} gives an overview of the package.
}
