% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.session.R
\name{taf.session}
\alias{taf.session}
\title{TAF Session}
\usage{
taf.session(sort = FALSE)
}
\arguments{
\item{sort}{whether to sort packages by name.}
}
\value{
Data from containing four columns:
\item{Package}{package name}
\item{Version}{package version}
\item{Library}{library where package was loaded from}
\item{RemoteSha}{GitHub reference code}
}
\description{
Show session information about loaded packages, clearly indicating which
packages were loaded from the local TAF library.
}
\examples{
\donttest{
taf.session()
taf.session(TRUE)
}

}
\seealso{
\code{\link{sessionInfo}} and the \pkg{sessioninfo} package provide further
session information about the version of R, etc.

\code{\link{taf.libPaths}} adds the TAF library to the search path for R
packages.

\code{\link{icesTAF-package}} gives an overview of the package.
}
