% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.skeleton.R
\name{taf.skeleton}
\alias{taf.skeleton}
\title{Create a Skeleton for a New TAF Analysis}
\usage{
taf.skeleton(name = "aTAFanalysis", list = character(),
  environment = .GlobalEnv, path = ".", force = FALSE,
  code_files = character(), encoding = "UTF-8-DOM")
}
\arguments{
\item{name}{character string: the package name and directory name for your package.}

\item{list}{character vector naming the R objects to put in the package. Usually, at most one of list, environment, or code_files will be supplied. See ‘Details’.}

\item{environment}{an environment where objects are looked for. See ‘Details’.}

\item{path}{path to put the package directory in.}

\item{force}{If FALSE will not overwrite an existing directory.}

\item{code_files}{a character vector with the paths to R code files to build the package around. See ‘Details’.}

\item{encoding}{optionally a character string with an encoding for an optional Encoding: line in ‘DESCRIPTION’ when non-ASCII characters will be used; typically one of "latin1", "latin2", or "UTF-8"; see the WRE manual.}
}
\description{
\code{taf.skeleton()} is motivated by \code{package.skeleton}. It automates
some of the setup for a new TAF analysis. It creates directories, saves functions,
data, and R code files to appropriate places, and creates a ‘Read-and-delete-me’
file describing further steps to create a functioning TAF analysis.
}
