% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.skeleton.sa.org.R
\name{taf.skeleton.sa.org}
\alias{taf.skeleton.sa.org}
\title{TAF Skeleton}
\usage{
taf.skeleton.sa.org(path = ".", stockname, force = FALSE)
}
\arguments{
\item{path}{where to create initial directories and R scripts. The default is
the current working directory.}

\item{stockname}{The short-form name of a stock on stockassessment.org.}

\item{force}{whether to overwrite existing scripts.}
}
\value{
Full path to analysis directory.
}
\description{
Create initial directories and R scripts for a new TAF analysis using
a stock assessment created on stockassessment.org.
}
\examples{
\dontrun{
taf.skeleton.sa.org(stockname = "WBCod_2021_cand01")
}

}
\seealso{
\code{\link{package.skeleton}} creates an empty template for a new R package.

\code{\link{TAF-package}} gives an overview of the package.
}
