% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.sources.R
\name{taf.sources}
\alias{taf.sources}
\title{Extract sources from TAF *.bib file}
\usage{
taf.sources(type)
}
\arguments{
\item{type}{one of "data", "software" or "both"}
}
\description{
Extract a list of sources from a TAF *.bib file (i.e. DATA.bib or
SOFTWARE.bib). This allows the user to print the lisst of sources
but also to process them individualy, giving more flexibiulity when
developing larger projects.
}
\seealso{
\link{process.entry} to process one of the entries returned by
`taf.sources`.
}
