% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.template.R
\name{taf.template}
\alias{taf.template}
\title{TAF Skeleton}
\usage{
taf.template(template = "WKTNET:simple", force = FALSE, gitignore = TRUE,
  path = ".")
}
\arguments{
\item{template}{which template to use. The default is "WKTNET:simple"}

\item{force}{whether to overwrite existing scripts.}

\item{gitignore}{whether to add a .gitignore file.}

\item{path}{where to create initial directories and R scripts. The default is
the current working directory.}
}
\value{
Full path to analysis directory.
}
\description{
Create initial directories and R scripts for a new TAF analysis using
a stock assessment created on stockassessment.org.
}
\examples{
\dontrun{
  taf.template(template = "WKTNET:simple")
}

}
\seealso{
\code{\link[TAF]{taf.skeleton}} creates an empty generic TAF skeleton.

\code{\link{icesTAF-package}} gives an overview of the package.
}
