% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.unix.R, R/taf.windows.R
\name{taf.unix}
\alias{taf.unix}
\alias{taf.windows}
\title{OS Family}
\usage{
taf.unix()

taf.windows()
}
\description{
Determine whether operating system is Windows or Unix-compatible.
}
\note{
This shorthand function can be useful when writing workaround solutions in
platform-independent scripts.
}
\examples{
taf.unix()
taf.windows()

}
\seealso{
\code{\link{.Platform}} reports the \code{OS.type} family.

\code{\link{icesTAF-package}} gives an overview of the package.
}
