% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf2long.R
\name{taf2long}
\alias{taf2long}
\title{Convert TAF table to long format}
\usage{
taf2long(x, names = c("Year", "Age", "Value"), year = TRUE)
}
\arguments{
\item{x}{a data frame containing crosstab data.}

\item{names}{a vector of three column names for the resulting data frame.}

\item{year}{whether the input table has year in the first column, instead of
row names.}
}
\value{
A data frame with three columns.
}
\description{
Convert a TAF crosstab table to long format.
}
\note{
TAF stores crosstab tables as data frames with a year column, as seen in
stock assessment reports. The long format is practical for analysis and
producing plots.
}
\examples{
fish <- matrix(round(runif(30),3), 5)
fish <- data.frame(Year=2001:2005, fish, check.names=FALSE)
taf2long(fish)

alt <- fish[-1]
row.names(alt) <- fish$Year
taf2long(alt, names=c("Year","Age","Weight"), year=FALSE)

}
\seealso{
\code{\link{flr2taf}} converts tables from FLR to TAF format.

\code{\link{icesTAF-package}} gives an overview of the package.
}
