% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf_api.R
\name{taf_api}
\alias{taf_api}
\title{Build a SAG web service url}
\usage{
taf_api(service, ...)
}
\arguments{
\item{service}{the name of the service}

\item{...}{name arguments will be added as queries}
}
\value{
a complete url as a character string
}
\description{
utility to build a url with optional query arguments
}
\examples{

taf_api("hi", bye = 21)
taf_api("FLStocks")

}
