% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf_post.R
\name{taf_post}
\alias{taf_post}
\title{Post to a url}
\usage{
taf_post(url, body = list(), verbose = FALSE)
}
\arguments{
\item{url}{the url to get.}

\item{body}{a list of named arguments to be sent as the body of the
post request.}

\item{verbose}{should verbose output form the http request be
returned? default FALSE.}
}
\value{
content of the http response.
}
\description{
Post to a url using an ICES authentication token
}
\seealso{
\code{\link{taf_api}} builds a TAF web service url.

\code{\link{icesTAF-package}} gives an overview of the package.
}
