% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taflib.R
\name{taflib}
\alias{taflib}
\title{TAF Library}
\usage{
taflib()
}
\value{
A vector of file paths.
}
\description{
Add TAF library to the search path where packages are stored.
}
\note{
This function inserts the entry \code{"bootstrap/library"} in front of
the existing library search path.
}
\examples{
.libPaths()
taflib()
.libPaths()

}
\seealso{
\code{\link{.libPaths}} gets/sets the library search path.

\code{\link{library}} looks for packages to load from the library search
 path.

\code{\link{icesTAF-package}} gives an overview of the package.
}
