% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.fay.R
\name{upload.fay}
\alias{upload.fay}
\title{Upload Fishing Mortality at Age}
\usage{
upload.fay(data, info, quiet = TRUE, only.check = FALSE, ...)
}
\arguments{
\item{data}{a dataframe with columns for year followed by columns
for each age, named as the age.}

\item{info}{a named list corresponding to the required
information to upload a stock assessment.}

\item{quiet}{should messages and verbose output be shown.}

\item{only.check}{logical, if TRUE, the data will be checked and
no upload will be attempted.}
}
\value{
TRUE if successfull, FALSE otherwise
}
\description{
Upload fishing mortality or harvest rate to the ICES TAF results
database
}
\details{
Upload a table of F at age to the ICES TAF results database.
The \code{data} and \code{assessment_info} will be checked against
a schema, and any errors reported back to the user as attributes to
the FALSE return value.
}
\note{
The \code{data} argument expects a data.frame with the first
column named "year" as described in the help for
\code{\link{xtab2taf}}.
}
\examples{
info <-
 list(
   valueType = "harvest",
   stockCode = "cod.27.47d20"
 )
upload.fay(catage.taf, info, only.check = TRUE)

}
\seealso{
The \code{\link{xtab2taf}} function is to create a data.frame

\code{\link{taf.png}} opens a PNG graphics device.

\code{\link{icesTAF-package}} gives an overview of the package.
}
