% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.artifact.R
\name{write.artifacts}
\alias{write.artifacts}
\title{Write an Artifact File}
\usage{
write.artifacts(..., append = TRUE, check = TRUE)
}
\arguments{
\item{...}{a list of artifact objects made using \code{artifact}.}

\item{append}{logical, should the new artifacts be appended to the
existing file?}

\item{check}{logical, should the artifact be checked for validity?}
}
\description{
Validate and write an artifact file.
}
