% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDLS.R
\name{writeDLS}
\alias{writeDLS}
\title{Write DLS Advice to File}
\usage{
writeDLS(x, file)
}
\arguments{
\item{x}{a list generated by the \code{DLS3.2} function.}

\item{file}{a filename.}
}
\description{
Write the advisory calculations for a data-limited stock (DLS) to a file.
}
\examples{
\dontrun{
library(icesAdvice)
survey <- data.frame(year=2001:2010, randu[1:10,])
dls <- DLS3.2(1000, survey$y)
writeDLS(dls, "dls-output.txt")
dls2 <- readDLS("dls-output.txt")
}
}
\seealso{
\code{\link{readDLS}} reads DLS advisory calculations from a file into a
list.

\code{\link[icesAdvice]{DLS3.2}} in the \pkg{icesAdvice} package can be used
to calculate catch advice for data-limited stocks.

\code{\link{icesTAF-package}} gives an overview of the package.
}

