% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab2long.R
\name{xtab2long}
\alias{xtab2long}
\title{Convert Crosstab Table to Long Format}
\usage{
xtab2long(x, names = c("Year", "Age", "Value"))
}
\arguments{
\item{x}{a data frame in crosstab format.}

\item{names}{a vector of three column names for the resulting data frame.}
}
\value{
A data frame with three columns.
}
\description{
Convert a table from crosstab format to long format.
}
\examples{
xtab2long(catage.xtab, names=c("Year","Age","Catch"))

}
\seealso{
\code{\link{catage.xtab}} and \code{\link{catage.long}} describe the crosstab
and long formats.

\code{\link{xtab2taf}} and \code{\link{taf2long}} are the underlying
functions that perform the conversion.

\code{\link{icesTAF-package}} gives an overview of the package.
}
