% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean TAF Directories}
\usage{
clean(path = ".")
}
\arguments{
\item{path}{location where directories are found.}
}
\description{
Remove directories \file{db}, \file{input}, \file{model}, \file{output}, and
\file{upload}.
}
\note{
The purpose of removing the directories is to make sure that subsequent TAF
scripts start by creating new empty directories.
}
\examples{
\dontrun{
clean()
}

}
\seealso{
\code{\link{sourceTAF}} runs a TAF script.

\code{\link{sourceAtoZ}} runs all TAF scripts in a directory.

\code{\link{icesTAF-package}} gives an overview of the package.
}
