% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean TAF Directories}
\usage{
clean(dirs = c("data", "input", "model", "output", "report"))
}
\arguments{
\item{dirs}{directories to delete.}
}
\description{
Remove TAF directories: \verb{data}, \verb{input}, \verb{model},
\verb{output}, \verb{report}.
}
\note{
The purpose of removing the directories is to make sure that subsequent TAF
scripts start by creating new empty directories.
}
\examples{
\dontrun{
clean()
}

}
\seealso{
\code{\link{sourceTAF}} runs a TAF script.

\code{\link{sourceAll}} runs all TAF scripts in a directory.

\code{\link{icesTAF-package}} gives an overview of the package.
}
