% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.R
\name{cp}
\alias{cp}
\title{Copy File}
\usage{
cp(from, to)
}
\arguments{
\item{from}{source file name.}

\item{to}{destination file name.}
}
\value{
\code{TRUE} for success, \code{FALSE} for failure, invisibly.
}
\description{
Copy file, overwriting existing file if necessary, and returning the result
invisibly.
}
\note{
Shorthand for \code{invisible(file.copy(..., overwrite=TRUE))}.
}
\examples{
\dontrun{
write(pi, "A.txt")
cp("A.txt", "B.txt")
file.remove("A.txt", "B.txt")
}

}
\seealso{
\code{\link{file.copy}} is the base function to copy files.

\code{\link{icesTAF-package}} gives an overview of the package.
}
