% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.remove.R
\name{dir.remove}
\alias{dir.remove}
\title{Remove Empty Directory}
\usage{
dir.remove(path)
}
\arguments{
\item{path}{a directory name.}
}
\value{
0 for success, 1 for failure, invisibly.
}
\description{
Remove an empty directory under any operating system.
}
\examples{
\dontrun{
dir.create("emptydir")
dir.remove("emptydir")
}

}
\seealso{
\code{\link{unlink}} can remove a non-empty directory.

\code{\link{dir.create}} creates an empty directory.

\code{\link{icesTAF-package}} gives an overview of the package.
}
