% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.remove.R
\name{dir.remove}
\alias{dir.remove}
\title{Remove Empty Directory}
\usage{
dir.remove(path, recursive = FALSE)
}
\arguments{
\item{path}{a directory name.}

\item{recursive}{whether to remove empty subdirectories as well.}
}
\value{
0 for success, 1 for failure, invisibly.
}
\description{
Remove empty directories under any operating system.
}
\note{
To tidy up a project directory, it can be useful to remove empty
subdirectories, while leaving non-empty directories intact.

The base function \code{unlink(dir, recursive=FALSE)} does not remove empty
directories in Windows and \code{unlink(dir, recursive=TRUE)} removes
non-empty directories, making it unsuitable for tidying up empty ones.
}
\examples{
\dontrun{
dir.create("emptydir")
dir.remove("emptydir")

dir.create("outer/inner", recursive=TRUE)
dir.remove("outer/inner", recursive=TRUE)
}

}
\seealso{
\code{\link{unlink}} can remove non-empty directories.

\code{\link{dir.create}} creates an empty directory.

\code{\link{icesTAF-package}} gives an overview of the package.
}
