% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.file.R, R/ds.package.R,
%   R/icesTAF-internal.R, R/process.inner.R, R/taf.install.R
\name{ds.file}
\alias{ds.file}
\alias{ds.package}
\alias{icesTAF-internal}
\alias{process.inner}
\alias{already.in.taf.library}
\title{Various TAF Helper Functions}
\usage{
ds.file(package, author, year, title, version, source)

ds.package(package, author, year, title, version, source)

process.inner(bib, dir, quiet)

already.in.taf.library(targz, lib)
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{icesTAF-package}} gives an overview of the package.
}
\keyword{internal}
