% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.github.R, R/ds.file.R, R/ds.package.R,
%   R/icesTAF-internal.R, R/parse.repo.R, R/process.inner.R, R/taf.install.R,
%   R/taf.libPaths.R, R/taf.session.R
\name{extract.subdir}
\alias{extract.subdir}
\alias{stamp.description}
\alias{ds.file}
\alias{ds.package}
\alias{icesTAF-internal}
\alias{parse.repo}
\alias{process.inner}
\alias{already.in.taf.library}
\alias{taf.libPaths}
\alias{taf.session}
\title{Various TAF Helper Functions}
\usage{
extract.subdir(targz, subtargz, subdir)

stamp.description(targz, spec, sha.full)

ds.file(package, author, year, title, version, source)

ds.package(package, author, year, title, version, source)

parse.repo(repo)

process.inner(bib, dir, quiet)

already.in.taf.library(targz, lib)

taf.libPaths()

taf.session(sort = TRUE)
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{icesTAF-package}} gives an overview of the package.
}
\keyword{internal}
