% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesTAF-package.R
\docType{package}
\name{icesTAF-package}
\alias{icesTAF-package}
\alias{icesTAF}
\title{Functions to Support the ICES Transparent Assessment Framework}
\description{
Functions to support the ICES Transparent Assessment Framework, to organize
data, methods, and results used in ICES assessments.
}
\details{
\emph{Initial TAF steps:}
\tabular{ll}{
  \code{\link{draft.data.script}} \tab create boot script\cr
  \code{\link{taf.roxygenise}}    \tab create DATA.bib entry from tags
}
}
\references{
ICES Transparent Assessment Framework: \url{https://taf.ices.dk}.

To explore example TAF stock assessments, see the introductory
\href{https://www.youtube.com/watch?v=FweJbr9hfdY}{video} and
\href{https://github.com/ices-taf/doc/blob/master/tutorial-1/README.md}{tutorial}.

The \href{https://github.com/ices-taf/doc/wiki}{TAF Wiki} provides additional
help resources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://taf.ices.dk}
  \item \url{https://github.com/ices-tools-prod/icesTAF}
}

}
\author{
Colin Millar and Arni Magnusson.
}
