% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesTAF-package.R
\docType{package}
\name{icesTAF-package}
\alias{icesTAF-package}
\alias{icesTAF}
\title{Functions to Support the ICES Transparent Assessment Framework}
\description{
Functions to support the ICES Transparent Assessment Framework, to organize
data, methods, and results used in ICES assessments.
}
\details{
\emph{Data-limited stocks:}
\tabular{ll}{
  \code{\link{readDLS}}  \tab read \code{DLS3.2} output from file\cr
  \code{\link{writeDLS}} \tab write \code{DLS3.2} output to file
}
\emph{Table converters:}
\tabular{ll}{
  \code{\link{flr2taf}}  \tab from FLR to TAF format\cr
  \code{\link{taf2long}} \tab from TAF to long format
}
\emph{File management:}
\tabular{ll}{
  \code{\link{dir.remove}} \tab remove empty directories\cr
  \code{\link{dos2unix}}   \tab convert line endings\cr
  \code{\link{unix2dos}}   \tab convert line endings
}
\emph{Examples:}
\tabular{ll}{
  \code{\link{catage.long}} \tab long format\cr
  \code{\link{catage.taf}}  \tab TAF format\cr
  \code{\link{catage.xtab}} \tab crosstab format
}
}
\references{
ICES Transparent Assessment Framework: \url{http://taf.ices.dk}.
}
\author{
Arni Magnusson and Colin Millar.
}
