% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin1.to.utf8.R, R/utf8.to-latin1.R
\name{latin1.to.utf8}
\alias{latin1.to.utf8}
\alias{utf8.to.latin1}
\title{Convert File Encoding}
\usage{
latin1.to.utf8(file, force = FALSE)

utf8.to.latin1(file, force = FALSE)
}
\arguments{
\item{file}{a filename.}

\item{force}{whether to perform the conversion even if the current file
encoding cannot be verified with \code{\link{file.encoding}}. Not
recommended.}
}
\description{
Convert file encoding between \verb{"latin1"} and \verb{"UTF-8"}.
}
\note{
In TAF, text files that have non-ASCII characters must be encoded as UTF-8.
}
\examples{
\dontrun{
utf8.to.latin1("data.txt")
latin1.to.utf8("data.txt")
}

}
\seealso{
\code{\link{iconv}} converts the encoding of a string.

\code{\link{file.encoding}} examines the encoding of a file.

\code{\link{icesTAF-package}} gives an overview of the package.
}
