% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd.R
\name{rnd}
\alias{rnd}
\title{Round Columns}
\usage{
rnd(x, cols, digits = 0, grep = FALSE)
}
\arguments{
\item{x}{a data frame.}

\item{cols}{column names, or column indices.}

\item{digits}{number of decimal places.}

\item{grep}{whether \code{cols} is a regular expression.}
}
\value{
A data frame similar to \code{x}, after rounding columns \code{cols} to the
number of \code{digits}.
}
\description{
Round column values in a data frame.
}
\note{
Provides notation that is reliable and convenient for modifying many columns
at once.
}
\examples{
# Equivalent alternatives:

summary <- rnd(summary.taf, c("Fbar", "Fbar_lo", "Fbar_hi"), 2)

summary <- rnd(summary.taf, "Fbar", 2, grep=TRUE)

}
\seealso{
\code{\link{round}} is the underlying function used to round numbers.

\code{\link{grep}} is the underlying function used to match column names if
\code{grep} is \code{TRUE}.

\code{\link{div}} is a similar function to divide columns with a common
number.

\code{\link{icesTAF-package}} gives an overview of the package.

The \pkg{icesAdvice} package provides the \code{\link[icesAdvice]{icesRound}}
function to round values for ICES advice sheets.
}
