% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceAll.R
\name{sourceAll}
\alias{sourceAll}
\title{Run TAF Scripts in Alphabetical Order}
\usage{
sourceAll(path = ".", rm = TRUE, clean = TRUE, quiet = FALSE)
}
\arguments{
\item{path}{directory containing TAF scripts.}

\item{rm}{whether to remove all objects from the global environment before
each script is run.}

\item{clean}{whether to \code{\link{clean}} existing TAF directories before
running the scripts.}

\item{quiet}{whether to suppress messages reporting progress.}
}
\value{
Logical vector, indicating which scripts ran without errors.
}
\description{
Run all TAF scripts in a directory, in alphabetical order. Optionally, start
by cleaning old output directories and emptying the workspace before running
each script.
}
\details{
By default, TAF scripts are run with \code{rm = TRUE} to make sure each
script starts with an empty workspace. Likewise, the default
\code{clean = TRUE} makes sure that the scripts start by creating new empty
directories and populate them one by one.
}
\examples{
\dontrun{
sourceAll()
}

}
\seealso{
\code{\link{sourceTAF}} runs a TAF script.

\code{\link{make}} runs a TAF script if needed.

\code{\link{clean}} cleans TAF directories.

\code{\link{icesTAF-package}} gives an overview of the package.
}
