% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceAll.R
\name{sourceAll}
\alias{sourceAll}
\title{Run All TAF Scripts}
\usage{
sourceAll(...)
}
\arguments{
\item{\dots}{passed to \code{\link{sourceTAF}}.}
}
\value{
Logical vector, indicating which scripts ran without errors.
}
\description{
Run core TAF scripts in current directory.
}
\note{
TAF scripts that will be run if they exist: \verb{data.R}, \verb{input.R},
\verb{model.R}, \verb{output.R}, and \verb{report.R}.

If a \verb{begin.R} script exists, it is ignored.
}
\examples{
\dontrun{
sourceAll()
}

}
\seealso{
\code{\link{sourceTAF}} runs a TAF script.

\code{\link{makeAll}} runs all TAF scripts as needed.

\code{\link{clean}} cleans TAF directories.

\code{\link{icesTAF-package}} gives an overview of the package.
}
