% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitchTAF.R
\name{stitchTAF}
\alias{stitchTAF}
\title{Stitch a TAF Script into an HTML report}
\usage{
stitchTAF(script, rm = TRUE, quiet = FALSE)
}
\arguments{
\item{script}{script filename.}

\item{rm}{whether to remove all objects from the global environment before
the script is stitched.}

\item{quiet}{whether to suppress messages reporting progress.

By default, TAF scripts are stitched with \code{rm = TRUE} to make sure each
script starts with an empty workspace.}
}
\value{
Invisible \code{TRUE} or \code{FALSE}, indicating whether the script ran
without errors.
}
\description{
Use the knitr function \code{stitch()} to generate an html report of a single
TAF script.
}
\note{
Commands within a script may change the working directory, but
\code{stitchTAF} guarantees that after running a script, the working
directory reported by \code{getwd()} is the same before and after running a
script.
}
\seealso{
\code{\link{stitch}} is the base function to stitch R scripts.

\code{\link{make}} runs a TAF script if needed.

\code{\link{sourceTAF}} runs a single TAF script.

\code{\link{sourceAll}} runs all TAF scripts in a directory.

\code{\link{icesTAF-package}} gives an overview of the package.
}
