% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.bootstrap.R
\name{taf.bootstrap}
\alias{taf.bootstrap}
\title{Bootstrap TAF Analysis}
\usage{
taf.bootstrap()
}
\description{
Set up data files and software required for the analysis. Model configuration
files are also set up, if found.
}
\note{
This function should be called from the top directory of a TAF analysis. It
looks for a directory called \file{bootstrap} containing metadata and
possibly other files.

The bootstrap procedure consists of the following steps:
\enumerate{
\item If a directory \verb{bootstrap/initial/config} contains model
configuration files, they are copied to \verb{bootstrap/config}.
\item If a \verb{bootstrap/DATA.bib} metadata file exists, it is processed.
\item If a \verb{bootstrap/SOFTWARE.bib} metadata file exists, it is
processed.
}

To override this default bootstrap procedure, the user can supply a custom
\verb{bootstrap.R} script that is run instead.
}
\examples{
\dontrun{
taf.bootstrap()
}

}
\seealso{
\code{\link{icesTAF-package}} gives an overview of the package.
}
