% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.bootstrap.R
\name{taf.bootstrap}
\alias{taf.bootstrap}
\title{Bootstrap TAF Analysis}
\usage{
taf.bootstrap(clean = TRUE, data = TRUE, software = TRUE,
  quiet = FALSE)
}
\arguments{
\item{clean}{whether to \code{\link{clean}} directories during the bootstrap
procedure.}

\item{data}{whether to process data files.}

\item{software}{whether to process software.}

\item{quiet}{whether to suppress messages reporting progress.}
}
\value{
Logical vector indicating which metadata files were processed.
}
\description{
Set up data files and software required for the analysis.
}
\note{
This function should be called from the top directory of a TAF analysis. It
looks for a directory called \file{bootstrap} and prepares data files and
software according to metadata specifications.

The bootstrap procedure consists of the following steps:
\enumerate{
\item If a \verb{bootstrap/DATA.bib} metadata file exists, it is processed
      with \code{\link{process.bib}}.
\item If a \verb{bootstrap/SOFTWARE.bib} metadata file exists, it is
      processed with \code{\link{process.bib}}.
}

After the bootstrap procedure, data and software have been documented and
are ready to be used in the subsequent analysis. Specifically, the procedure
populates up to four new directories:
\itemize{
\item \verb{bootstrap/data} with data files.
\item \verb{bootstrap/library} with R packages compiled for the local
      platform.
\item \verb{bootstrap/software} with software files, such as R packages in
      \verb{tar.gz} source code format.
}

Model settings and configuration files can be set up within
\verb{SOFTWARE.bib}, see \code{\link{process.bib}}.
}
\examples{
\dontrun{
taf.bootstrap()
}

}
\seealso{
\code{\link{process.bib}} is a helper function used to process metadata.

\code{\link{taf.library}} loads a package from \verb{bootstrap/library}.

\code{\link{icesTAF-package}} gives an overview of the package.
}
