% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.bootstrap.R
\name{taf.bootstrap}
\alias{taf.bootstrap}
\title{Bootstrap TAF Analysis}
\usage{
taf.bootstrap(software = TRUE, data = TRUE, clean = TRUE,
  force = FALSE, taf = NULL, quiet = FALSE)
}
\arguments{
\item{software}{whether to process \verb{SOFTWARE.bib}.}

\item{data}{whether to process \verb{DATA.bib}.}

\item{clean}{whether to \code{\link{clean}} directories during the bootstrap
procedure.}

\item{force}{whether to remove existing \verb{bootstrap/data},
\verb{bootstrap/library}, and \verb{bootstrap/software} directories
before the bootstrap procedure.}

\item{taf}{a convenience flag where \code{taf = TRUE} sets \code{software},
\code{data}, \code{clean}, and \code{force} to \code{TRUE}, as is done
on the TAF server. Any other value of \code{taf} is ignored.}

\item{quiet}{whether to suppress messages reporting progress.}
}
\value{
Logical vector indicating which metadata files were processed.
}
\description{
Process metadata files \file{SOFTWARE.bib} and \file{DATA.bib} to set up
software and data files required for the analysis.
}
\details{
If \code{clean = TRUE} then:
\enumerate{
\item \code{\link{clean.software}} and \code{\link{clean.library}} are run if
      \file{SOFTWARE.bib} is processed.
\item \code{\link{clean.data}} is run if \file{DATA.bib} is processed.
}

The default behavior of \code{taf.bootstrap} is to skip downloading of remote
files (GitHub resources, URLs, bootstrap scripts) and also skip installing R
packages from GitHub if the files seem to be already in place. This is done
to speed up the bootstrap procedure as much as possible. To override this and
guarantee that all data and software files are updated, pass \code{force =
TRUE} to download and install everything declared in \verb{SOFTWARE.bib} and
\verb{DATA.bib}.
}
\note{
This function should be called from the top directory of a TAF analysis. It
looks for a directory called \file{bootstrap} and prepares data files and
software according to metadata specifications.

The bootstrap procedure consists of the following steps:
\enumerate{
\item If a \verb{bootstrap/SOFTWARE.bib} metadata file exists, it is
      processed.
\item If a \verb{bootstrap/DATA.bib} metadata file exists, it is processed.
}

After the bootstrap procedure, software and data have been documented and
are ready to be used in the subsequent analysis. Specifically, the procedure
populates up to three new directories:
\itemize{
\item \verb{bootstrap/data} with data files.
\item \verb{bootstrap/library} with R packages compiled for the local
      platform.
\item \verb{bootstrap/software} with software files, such as R packages in
      \verb{tar.gz} source code format.
}

Model settings and configuration files can be set up within \verb{DATA.bib},
see \href{https://github.com/ices-taf/doc/wiki/Bib-entries}{TAF Wiki}.
}
\examples{
\dontrun{
taf.bootstrap()
}

}
\seealso{
\code{\link{draft.data}} and \code{\link{draft.software}} can be used to
create initial draft versions of \file{DATA.bib} and \file{SOFTWARE.bib}
metadata files.

\code{\link{taf.library}} loads a package from \verb{bootstrap/library}.

\code{\link{icesTAF-package}} gives an overview of the package.
}
