% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.colors.R
\docType{data}
\name{taf.colors}
\alias{taf.colors}
\alias{taf.green}
\alias{taf.orange}
\alias{taf.blue}
\alias{taf.dark}
\alias{taf.light}
\title{TAF Colors}
\usage{
taf.green
taf.orange
taf.blue
taf.dark
taf.light
}
\description{
Predefined colors for TAF plots.
}
\examples{
taf.green
barplot(5:1, col=c(taf.green, taf.orange, taf.blue, taf.dark, taf.light))
}
\seealso{
\code{\link{icesTAF-package}} gives an overview of the package.
}
