% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.install.R
\name{taf.install}
\alias{taf.install}
\title{Install Package from GitHub into TAF Library}
\usage{
taf.install(owner, repo, ref, subdir = "")
}
\arguments{
\item{owner}{GitHub user name or organization hosting the repository.}

\item{repo}{GitHub repository name.}

\item{ref}{reference, preferably a tag, release, or SHA-1 hash.}

\item{subdir}{subdirectory containing the R package, if it's not at the top
level of the repository.}
}
\description{
Install an R package that is hosted on GitHub. The package is installed into
a local TAF library \file{bootstrap/library} within the assessment directory
tree.
}
\note{
A fixed reference such as a tag, release, or SHA-1 hash is recommended.
Branch names, such as \code{"master"}, are pointers that are subject to
change, and are therefore not reliable as long-term references.

This function is not intended as a replacement for the \code{install_github}
function found in the \pkg{devtools} package. Rather, it serves as a
reference tool to support long-term reproducibility of TAF assessments,
without adding \pkg{devtools} and its underlying packages as dependencies.

The purpose of the TAF library is to retain R packages used in a TAF
assessment that are not archived on CRAN.
}
\examples{
\dontrun{
# Installing the tagged release 1.0-0
taf.install("ices-tools-prod", "icesTAF", "1.0-0")

# is equivalent to the 40-character SHA-1 hash
taf.install("ices-tools-prod", "icesTAF",
            "69beeb4300453d93a10a5a35142775119df7da66")

# SHA-1 hashes are commonly shortened to 7 characters
taf.install("ices-tools-prod", "icesTAF", "69beeb4")
}

}
\seealso{
\code{\link{taf.library}} adds the TAF library to the search path.

\code{\link{install.packages}} is the underlying base function that installs
the package after downloading from GitHub.

\code{\link[devtools]{install_github}} is a commonly used function to install
packages from GitHub.

\code{\link{icesTAF-package}} gives an overview of the package.
}
