% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.png.R
\name{taf.png}
\alias{taf.png}
\alias{tafpng}
\title{PNG Device}
\usage{
taf.png(filename, width = 800, height = 600, pointsize = 22, ...)
}
\arguments{
\item{filename}{image filename.}

\item{width}{image width.}

\item{height}{image height.}

\item{pointsize}{text size.}

\item{\dots}{passed to \code{png}.}
}
\description{
Open PNG graphics device to create an image file inside the TAF \verb{report}
folder.
}
\details{
The \code{filename} can be passed without the preceding
\code{"report/"}, and without the \code{".png"} filename extension.

Specifically, the function prepends \code{"report/"} to the filename if (1)
the filename does not contain a \code{"/"} separator, (2) the working
directory is not \verb{report}, and (3) the directory \verb{report} exists.
The function also appends \verb{".png"} to the filename if it does not
already have that filename extension.

This automatic filename manipulation can be bypassed by using the \code{png}
function directly.
}
\note{
A simple convenience function to shorten
\preformatted{png("report/myplot.png", width=800, height=600, pointsize=22)}
to
\preformatted{taf.png("myplot")}
To successfully export \verb{trellis} and \verb{ggplot} objects to image
files in scripts, these objects should be explicitly printed (see examples
below).
}
\examples{
\dontrun{
taf.png("myplot")
plot(1)
dev.off()

library(lattice)
taf.png("mytrellis")
print(xyplot(1~1))
dev.off()

library(ggplot2)
taf.png("myggplot")
print(qplot(1))
dev.off()
}

}
\seealso{
\code{\link{png}} is the underlying function used to open a PNG graphics
device.

\code{\link{icesTAF-package}} gives an overview of the package.
}
