% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.roxygenise.R
\name{taf.roxygenise}
\alias{taf.roxygenise}
\title{Process a TAF repo with the taf roclet}
\usage{
taf.roxygenise(path = ".", files)
}
\arguments{
\item{path}{location of taf repository top level directory. Default
is working directory.}

\item{files}{a vector of file names to parse for documentation.}
}
\description{
This function builds documentation for a TAF repository
using roxygen syntax headers.  It depnends on the roxygen2
package adding some extra functionality to produce citation
entries for data sources
}
