% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.skeleton.R
\name{taf.skeleton}
\alias{taf.skeleton}
\title{Create a Skeleton for a New TAF Analysis}
\usage{
taf.skeleton(name = "aTAFanalysis", path = ".", force = FALSE)
}
\arguments{
\item{name}{character string: the name and directory name for your analysis,
if NULL then it effectively assumes the parent directory name as the
analysis name.}

\item{path}{path to put the analysis directory in.}

\item{force}{If FALSE will not overwrite an existing directory.}
}
\description{
\code{taf.skeleton()} is motivated by \code{package.skeleton}. It automates
some of the setup for a new TAF analysis. Currently it creates directories and
empty R scripts with appropriate header information.
}
