% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.skeleton.R
\name{taf.skeleton}
\alias{taf.skeleton}
\title{TAF Skeleton}
\usage{
taf.skeleton(name = "analysis", path = ".", force = FALSE)
}
\arguments{
\item{name}{main directory name for the analysis.}

\item{path}{path to create the analysis directory in.}

\item{force}{whether to overwrite an existing directory.}
}
\value{
Full path to analysis directory.
}
\description{
Create an empty template for a new TAF analysis: initial directories and R
scripts with TAF header comments.
}
\details{
Use \code{name = "."} to create initial directories and scripts inside the
current working directory.
}
\examples{
\dontrun{
taf.skeleton()
}

}
\seealso{
\code{\link{package.skeleton}} creates an empty template for a new R package.

\code{\link{icesTAF-package}} gives an overview of the package.
}
