% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\name{tt}
\alias{tt}
\title{TAF Transpose}
\usage{
tt(x)
}
\arguments{
\item{x}{a data frame in TAF format.}
}
\value{
A data frame with years as column names.
}
\description{
Convert a table from TAF format to transposed crosstab format.
}
\note{
Transposing can be useful when comparing TAF tables to stock assessment
reports.
}
\examples{
taf2xtab(catage.taf)
tt(catage.taf)

}
\seealso{
\code{\link{catage.taf}} describes the TAF format.

\code{\link{taf2xtab}} converts a TAF table to crosstab format, without
transposing.

\code{\link{icesTAF-package}} gives an overview of the package.
}
