% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.taf.R
\name{write.taf}
\alias{write.taf}
\title{Write TAF Table to File}
\usage{
write.taf(x, file)
}
\arguments{
\item{x}{a data frame in TAF format.}

\item{file}{a filename.}
}
\description{
Write a TAF table to a file.
}
\examples{
\dontrun{
write.taf(catage.taf, "catage.csv")
catage <- read.taf("catage.csv")

file.remove("catage.csv")
}

}
\seealso{
\code{\link{read.taf}} reads a TAF table from a file into a data frame.

\code{\link{icesTAF-package}} gives an overview of the package.
}
